tidy a few Clang++ -Weverything warnings

+++ src/osmdata-sp.cpp
@@ -248,8 +248,8 @@ void get_osm_ways_sp (Rcpp::S4 &sp_ways,
         // Fill plotOrder slot with int vector - this has to be int, not
         // unsigned int!
         std::vector <int> plord;
-        for (int i=0; i<nrow; i++)
-            plord.push_back (i + 1);
+        for (size_t i=0; i<nrow; i++)
+            plord.push_back (static_cast <int> (i) + 1);
         sp_ways.slot ("plotOrder") = plord;
         plord.clear ();
         sp_ways.slot ("data") = kv_df;
